hal/riscv: apply new structure of hal interface
more:
 - code with the old syspage is temporarily disabled

JIRA: PD-21
