+++ hal/riscv64/Makefile
@@ -4,7 +4,8 @@
 # Copyright 2018, 2020 Phoenix Systems
 #
 
-OBJS += $(addprefix $(PREFIX_O)hal/riscv64/, _init.o string.o _interrupts.o syspage.o hal.o spinlock.o interrupts.o cpu.o pmap.o dtb.o timer.o exceptions.o plic.o)
+OBJS += $(addprefix $(PREFIX_O)hal/riscv64/, _init.o _string.o _interrupts.o hal.o spinlock.o interrupts.o cpu.o pmap.o dtb.o timer.o string.o exceptions.o plic.o)
+CFLAGS += -Ihal/riscv64
 
 OBJS += $(PREFIX_O)hal/riscv64/console.o
