+++ hal/riscv64/_init.S
@@ -30,8 +30,8 @@
 #define SR_SUM          0x00040000
 #define SR_MER          0x00080000
 
-#include "cpu.h"
-#include "pmap.h"
+#include <arch/cpu.h>
+#include <arch/pmap.h>
 
 
 .section .init, "x"
