+++ hal/riscv64/arch/interrupts.h
@@ -14,17 +14,13 @@
  * %LICENSE%
  */
 
-#ifndef _HAL_INTERRUPTS_H_
-#define _HAL_INTERRUPTS_H_
-
-#ifndef __ASSEMBLY__
+#ifndef _HAL_RISCV64_INTERRUPTS_H_
+#define _HAL_RISCV64_INTERRUPTS_H_
 
 #include "cpu.h"
-#include "pmap.h"
 
 
-#define SYSTICK_IRQ		0
-
+#define SYSTICK_IRQ 0
 
 typedef struct _intr_handler_t {
 	struct _intr_handler_t *next;
@@ -33,22 +29,5 @@ typedef struct _intr_handler_t {
 	int (*f)(unsigned int, cpu_context_t *, void *);
 	void *data;
 } intr_handler_t;
-
-
-/* Function installs new handler for interrupt given by n */
-extern int hal_interruptsSetHandler(intr_handler_t *h);
-
-
-extern int hal_interruptsDeleteHandler(intr_handler_t *h);
-
-
-extern char *hal_interruptsFeatures(char *features, unsigned int len);
-
-
-/* Function initializes interrupt handling */
-extern void _hal_interruptsInit(void);
-
-
-#endif
 
 #endif
