+++ hal/riscv64/arch/spinlock.h
@@ -1,0 +1,37 @@
+/*
+ * Phoenix-RTOS
+ *
+ * Operating system kernel
+ *
+ * Spinlock
+ *
+ * Copyright 2012, 2018, 2020 Phoenix Systems
+ * Author: Pawel Pisarczyk
+ *
+ * This file is part of Phoenix-RTOS.
+ *
+ * %LICENSE%
+ */
+
+#ifndef _HAL_RISCV64_SPINLOCK_H_
+#define _HAL_RISCV64_SPINLOCK_H_
+
+#include "types.h"
+
+
+typedef u64 spinlock_ctx_t;
+
+typedef struct _spinlock_t {
+	const char *name;
+	cycles_t b;
+	cycles_t e;
+	cycles_t dmin;
+	cycles_t dmax;
+	struct _spinlock_t *next;
+	struct _spinlock_t *prev;
+
+	u64 lock;
+	u64 sstatus;
+} spinlock_t;
+
+#endif
