+++ hal/riscv64/hal.c
@@ -13,18 +13,39 @@
  * %LICENSE%
  */
 
-#include "spinlock.h"
-#include "console.h"
-#include "exceptions.h"
-#include "interrupts.h"
-#include "cpu.h"
-#include "pmap.h"
-#include "timer.h"
+#include "../hal.h"
+#include "../spinlock.h"
+#include "../console.h"
+#include "../exceptions.h"
+#include "../interrupts.h"
+#include "../cpu.h"
+#include "../pmap.h"
+#include "../timer.h"
+#include "config.h"
 
 
 struct {
 	int started;
 } hal_common;
+
+hal_syspage_t *syspage;
+
+
+void *hal_syspageRelocate(void *data)
+{
+	return data;
+}
+
+
+ptr_t hal_syspageAddr(void)
+{
+	return (ptr_t)syspage;
+}
+
+
+void hal_wdgReload(void)
+{
+}
 
 
 int hal_started(void)
@@ -47,9 +68,12 @@ __attribute__ ((section (".init"))) void _hal_init(void)
 	_hal_exceptionsInit();
 	_hal_interruptsInit();
 
-	_timer_init(SYSTICK_INTERVAL);
-/*	_hal_cpuInit();*/
+	_hal_timerInit(SYSTICK_INTERVAL);
 
+#if 0
+	_hal_cpuInit();
+#endif
 	hal_common.started = 0;
+
 	return;
 }
