+++ hal/riscv64/spinlock.c
@@ -13,8 +13,8 @@
  * %LICENSE%
  */
 
-#include "spinlock.h"
-#include "cpu.h"
+#include "../spinlock.h"
+#include "../cpu.h"
 
 
 struct {
@@ -42,6 +42,51 @@ static void _hal_spinlockCreate(spinlock_t *spinlock, const char *name)
 		spinlock->next = spinlock;
 		spinlock->prev = spinlock;
 	}
+	return;
+}
+
+
+void hal_spinlockSet(spinlock_t *spinlock, spinlock_ctx_t *sc)
+{
+	__asm__ volatile
+	(" \
+		csrrc t0, sstatus, 2; \
+		sd t0, %0; \
+		mv t0, zero; \
+1: \
+		amoswap.w.aq t0, t0, %1; \
+		beqz t0, 1b"
+	:
+	: "m" (spinlock->sstatus), "A" (spinlock->lock)
+	: "t0", "memory");
+
+	/* (MOD) inefficiency because of rdcycles instruction */
+	/* hal_cpuGetCycles((void *)&spinlock->b); */
+}
+
+
+void hal_spinlockClear(spinlock_t *spinlock, spinlock_ctx_t *sc)
+{
+	/* (MOD) inefficiency because of rdcycles instruction */
+	/* hal_cpuGetCycles((void *)&spinlock->e);*/
+
+	/* Calculate maximum and minimum lock time */
+	/*if ((cycles_t)(spinlock->e - spinlock->b) > spinlock->dmax)
+		spinlock->dmax = spinlock->e - spinlock->b;
+
+	if (spinlock->e - spinlock->b < spinlock->dmin)
+		spinlock->dmin = spinlock->e - spinlock->b;*/
+
+	__asm__ volatile
+	(" \
+		li t1, 1; \
+		amoswap.w.rl t1, t1, %0; \
+		ld t0, %1; \
+		csrw sstatus, t0"
+	:
+	: "A" (spinlock->lock), "m" (spinlock->sstatus)
+	: "t0", "t1", "memory");
+
 	return;
 }
