--- hal/riscv64/syspage.c
@@ -1,20 +1,0 @@
-/*
- * Phoenix-RTOS
- *
- * Operating system kernel
- *
- * System information page (prepared by kernel loader)
- *
- * Copyright 2018 Phoenix Systems
- * Author: Pawel Pisarczyk
- *
- * This file is part of Phoenix-RTOS.
- *
- * %LICENSE%
- */
-
-#include "syspage.h"
-
-
-/* Syspage */
-syspage_t * const syspage;
